Let's split the module with auxiliary functions into two: general and utilities for tasks
